pyproject.toml
[tool.poetry]
name = "goit-web-hw-module-11"
version = "0.1.0"
description = ""
authors = ["Deminform <deminform@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.16"
psycopg2 = "^2.9.10"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
starlette = "^0.41.2"
python-jose = "^3.3.0"
libgravatar = "^1.0.4"
passlib = "^1.7.4"
bcrypt = "4.0.1"
fastapi-mail = "^1.4.1"
redis = "^5.2.0"
fastapi-limiter = "^0.1.6"
jinja2 = "^3.1.4"
cloudinary = "^1.41.0"
aioredis = "^2.0.1"
fastapi-cache2 = "^0.2.2"
flake8 = "^7.1.1"
flake8-async = "^24.11.3"
faker = "^33.1.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest = "^8.3.4"
httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
#addopts = "--doctest-modules"
#testpath = "tests"
asyncio_default_fixture_loop_scope = "session"
pythonpath = "."
filterwarnings = [
    "ignore:.*datetime\\.datetime\\.utcnow.*:DeprecationWarning"
]